dataManager: added new function updateDataObject, dataObjAnalysis: added new function updateDataObjectAnalysis
updateDataObject: this allow to substitute a dataObj with a new one. It is ment to update dataObj with results of analysis performed on the dataset. It is not meant to be used by users because it can mess up things pretty badly. In the future I may think to an alternative to make this function not accessible for user or something similar.
-updateDataObjectAnalysis: update dataObj analysis with a result